fix!: replace the custom Locale enum with the Locale java object
The carbonio-user-management changed the format of the returned locale
value of an account. Now it uses the standard format xx_YY, so the SDK
is aligned to map the returned locale value with the default java Locale
object instead of using a custom enum.

BREAKING CHANGE: If you are using the Locale enum on your codebase be
aware that is deleted from the SDK version 0.2.3.

refs: UM-25